+++ tests/testthat/test_Dataset_classes.R
@@ -22,6 +22,11 @@ test_that("RawDataset basics", {
   )
 
   expect_equal(
+    test_ds$get_column_labels(),
+    setNames(c("X", "Y"), c("x", "y"))
+  )
+
+  expect_equal(
     test_ds$get_character_colnames(),
     character(0)
   )
@@ -68,6 +73,11 @@ test_that("RawDataset basics", {
   })
 
   expect_equal(
+    RawDataset$new(x),
+    raw_dataset(x)
+  )
+
+  expect_equal(
     test_ds$get_character_colnames(),
     "y"
   )
@@ -81,6 +91,27 @@ test_that("RawDataset basics", {
     test_ds$get_numeric_colnames(),
     "x"
   )
+
+  expect_equal(
+    test_ds$get_colnames(),
+    test_ds$colnames
+  )
+
+  expect_equal(
+    test_ds$get_rownames(),
+    test_ds$rownames
+  )
+
+  expect_equal(
+    test_ds$get_colnames(),
+    test_ds$var_names
+  )
+
+  expect_equal(
+    get_raw_data(test_ds),
+    x
+  )
+
 
 })
 
@@ -101,7 +132,15 @@ test_that("NamedDataset basics", {
   )
 
   expect_silent(
-    NamedDataset$new(x = x, name = "abc")
+    NamedDataset$new(x = x, dataname = "abc")
+  )
+
+  expect_silent(
+    NamedDataset$new(x = x, dataname = "abc", label = NULL)
+  )
+  expect_equal(
+    NamedDataset$new(x = x, dataname = "abc", label = NULL),
+    named_dataset(x = x, dataname = "abc", label = NULL)
   )
 
   expect_error(
@@ -109,7 +148,7 @@ test_that("NamedDataset basics", {
   )
 
   expect_silent({
-    test_ds <- NamedDataset$new(x, name = "testds",
+    test_ds <- NamedDataset$new(x, dataname = "testds",
                                 code = "test_ds <- data.frame(x = c(1, 1), y = c('a', 'a'), stringsAsFactors = TRUE)",
                                 label = "Testing Dataset"
                                 )
@@ -123,6 +162,11 @@ test_that("NamedDataset basics", {
   expect_equal(
     "Testing Dataset",
     test_ds$dataset_label
+  )
+
+  expect_equal(
+    "Testing Dataset",
+    test_ds$get_dataset_label()
   )
 
   expect_equal(
@@ -145,23 +189,58 @@ test_that("RelationalDataset basics", {
   rtables::var_labels(x) <- c("X", "Y")
 
   expect_error(
-    RelationalDataset$new(x = x, name = "abc")
+    RelationalDataset$new(x = x, dataname = "abc")
   )
 
   expect_silent({
-    test_ds <- RelationalDataset$new(x, name = "testds",
-                                keys = list(primary = "x")
+    test_ds <- RelationalDataset$new(x, dataname = "testds",
+                                keys = keys(primary = "x", foreign = NULL, parent = NULL)
     )
   })
 
   expect_equal(
     test_ds$keys,
-    list(primary = "x")
+    keys(primary = "x", foreign = NULL, parent = NULL)
   )
 
-  expect_silent(test_ds$set_keys(list(primary = "y")))
+  expect_silent(test_ds$set_keys(keys(primary = "y", foreign = NULL, parent = NULL)))
   expect_equal(
     test_ds$keys,
-    list(primary = "y")
+    keys(primary = "y", foreign = NULL, parent = NULL)
+  )
+})
+
+## as_relational ====
+context("as_relational")
+
+test_that("as_relational function", {
+  x <- data.frame(x = c(1, 1), y = c("a", "a"), stringsAsFactors = TRUE)
+  rtables::var_labels(x) <- c("X", "Y")
+
+  expect_silent({
+    test_ds <- RawDataset$new(x)
+  })
+
+  expect_error({
+    as_relational(list(data = "a"))
+    },
+    "RawDataset"
+  )
+
+  expect_equal(
+    RelationalDataset$new(
+      x = x,
+      dataname = "abc",
+      keys = keys(primary = "x", foreign = NULL, parent = NULL),
+      code = "xx",
+      label = character(0)
+    ),
+    as_relational(
+      dataset = test_ds,
+      dataname = "abc",
+      keys = keys(primary = "x", foreign = NULL, parent = NULL),
+      code = "xx",
+      label = character(0)
+    )
   )
 })