+++ tests/testthat/test_mutate_dataset.R
@@ -1,0 +1,117 @@
+## mutate_dataset ====
+context("mutate_dataset")
+
+test_that("mutate_dataset", {
+
+  x <- data.frame(x = c(1, 1), y = c("a", "a"), stringsAsFactors = FALSE)
+
+  expect_silent({
+    test_ds <- RawDataset$new(x)
+  })
+
+  expect_error({
+    mutate_dataset(dataset = test_ds, code = "a")
+  }, "is")
+
+  expect_silent({
+    test_ds <- NamedDataset$new(x, dataname = "x",
+                                code = "data.frame(x = c(1, 1), y = c('a', 'a'), stringsAsFactors = TRUE)")
+  })
+
+  expect_error({
+    mutate_dataset(dataset = test_ds)
+  }, "is_character_vector")
+
+  expect_error({
+    mutate_dataset(dataset = test_ds, code = TRUE)
+  }, "character")
+
+  expect_error({
+    mutate_dataset(dataset = test_ds, code = "y <- test")
+  }, "use the dataname")
+
+  expect_silent({
+    test_ds_mut <- test_ds %>% mutate_dataset("x$z <- c('one', 'two')")
+  })
+
+  expect_equal(
+    test_ds_mut$get_raw_data(),
+    data.frame(x = c(1, 1), y = c("a", "a"),
+               z = c("one", "two"),
+               stringsAsFactors = FALSE)
+  )
+
+  expect_error({
+    test_ds_mut <- test_ds %>% mutate_dataset("x <- 3")
+  }, "data.frame")
+
+  expect_error({
+    test_ds_mut <- test_ds %>% mutate_dataset(c("x <- 3", "som"))
+  }, "max_length")
+
+  expect_silent({
+    test_ds <- RelationalDataset$new(x, dataname = "x",
+                                     keys = keys(primary = "x", foreign = NULL, parent = NULL)
+    )
+  })
+  expect_error({
+    test_ds_mut <- test_ds %>% mutate_dataset("testds$z <- c('one', 'two')")
+  })
+
+  expect_silent({
+    test_ds <- RelationalDataset$new(x, dataname = "testds", code = "testds <- whatever",
+                                     keys = keys(primary = "x", foreign = NULL, parent = NULL)
+    )
+  })
+
+  expect_silent({
+    test_ds_mut <- test_ds %>% mutate_dataset("testds$z <- c('one', 'two')")
+  })
+
+  expect_equal(
+    test_ds_mut$get_raw_data(),
+    data.frame(x = c(1, 1), y = c("a", "a"),
+               z = c("one", "two"),
+               stringsAsFactors = FALSE)
+  )
+
+  expect_silent({
+    test_ds_mut <- test_ds %>% mutate_dataset(read_script("mutate_code/testds.R"))
+  })
+
+  expect_equal(
+    test_ds_mut$get_raw_data(),
+    data.frame(x = c(1, 1), y = c("a", "a"),
+               z = c(1, 1),
+               stringsAsFactors = FALSE)
+  )
+
+  expect_equal(
+    test_ds_mut$get_code(),
+    "testds <- whatever\n\nmut_fun <- function(x){\n  x$z <- 1\n  return(x)\n}\ntestds <- mut_fun(testds)"
+  )
+
+  expect_true(is(test_ds_mut, "RelationalDataset"))
+
+  expect_silent({
+    test_ds_mut <- test_ds %>% mutate_dataset(script = "mutate_code/testds.R")
+  })
+
+  expect_equal(
+    test_ds_mut$get_raw_data(),
+    data.frame(x = c(1, 1), y = c("a", "a"),
+               z = c(1, 1),
+               stringsAsFactors = FALSE)
+  )
+
+  expect_equal(
+    test_ds_mut$get_code(),
+    "testds <- whatever\n\nmut_fun <- function(x){\n  x$z <- 1\n  return(x)\n}\ntestds <- mut_fun(testds)"
+  )
+
+  expect_true(is(test_ds_mut, "RelationalDataset"))
+
+  expect_error({
+    test_ds_mut <- test_ds %>% mutate_dataset(code = "rm('testds')")
+  }, "Mutations")
+})